Update theme.css
Dashboard selector switch alignment #50
I added the overflow tag to avoid the overlapping of the switches and the icons. It looks better but icon is still overlapped if the switch is not Highlighted. It looks nicer if user wants add more selectors. 
I solved also an alignment issue in switches with the switches selectors.